add test showing issue
